# docs/api/new-api-asset-assessment.openapi.yaml
openapi: 3.0.3
info:
  title: new-api Asset Pre-evaluation and Assessment API
  version: 1.0.0
  description: |
    Browser-facing contract implemented by new-api. Downstream GPUFabric and
    asset-assessment-service internal APIs are documented separately.
servers:
  - url: /
...
paths:
...
  /api/banking/provider/pre-evaluations/{taskRef}/pdf:
    parameters:
      - $ref: '#/components/parameters/TaskRef'
    get:
      tags: [Banking Pre-evaluations]
      summary: Get a verified PDF pre-evaluation report
      operationId: getPreEvaluationPdf
      parameters:
        - $ref: '#/components/parameters/NewApiUser'
      responses:
        '200':
          description: Rendered PDF report
          headers:
            Cache-Control:
              schema: {type: string, example: 'private, no-store'}
            X-Content-SHA256:
              schema: {$ref: '#/components/schemas/Sha256'}
          content:
            application/pdf:
              schema: {type: string, format: binary}
        '404': {$ref: '#/components/responses/NotFound'}
        '502': {$ref: '#/components/responses/BadGateway'}
        '503': {$ref: '#/components/responses/Unavailable'}
...
security:
  - UserBearer: []
  - UserSession: []
components:
  securitySchemes:
    UserBearer:
      type: http
      scheme: bearer
      description: Existing new-api user access token. New-Api-User is also required.
    UserSession:
      type: apiKey
      in: cookie
      name: session
      description: Existing new-api authenticated session cookie.
  parameters:
...
    NewApiUser:
      name: New-Api-User
      in: header
      required: true
      description: Required by existing new-api UserAuth middleware for both session and Bearer authentication.
      schema: {type: string}
...
    TaskRef:
      name: taskRef
      in: path
      required: true
      schema: {type: string, minLength: 8, maxLength: 64}
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [success, code, message, retryable, requestId]
      properties:
        success: {type: boolean, enum: [false]}
        code: {type: string}
        message: {type: string}
        retryable: {type: boolean}
        details: {type: object, additionalProperties: true}
        requestId: {type: string}
...
    Sha256:
      type: string
      pattern: '^[0-9a-f]{64}$'
  responses:
...
    NotFound:
      description: Resource does not exist or is not owned by the caller
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
    BadGateway:
      description: Upstream response or integrity validation failed
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    Unavailable:
      description: Required upstream or security configuration is unavailable
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...